# Fix n+1 query issue on link expansion.